+++ system/Classes/Worker/Application.php
@@ -2,6 +2,9 @@
 
 namespace Asylamba\Classes\Worker;
 
+use Asylamba\Classes\Configuration\Configuration;
+use Symfony\Component\Config\FileLocator;
+
 class Application {
     /** @var Container **/
     protected $container;
@@ -9,7 +12,18 @@ class Application {
     public function boot()
     {
         $this->container = new Container();
+		$this->configure();
     }
+	
+	public function configure()
+	{
+		$configurationFiles = [
+			__DIR__ . '/../../../config/parameters.yml',
+			__DIR__ . '/../../../config/services.yml'
+		];
+		$configuration = new Configuration(new FileLocator($configurationFiles));
+		$configuration->buildContainer($this->container, $configurationFiles);
+	}
     
     /**
      * @return Container